cleaned up some refrences and fixed an issue in the delete api

+++ InterviewTeaChallenge/JO.InterviewTeaChallenge.Core/Services/TeaService.cs
@@ -33,6 +33,8 @@ namespace JO.InterviewTeaChallenge.Core.Services
         {
             var tea = await _teaRepository.GetById(id);
 
+            if (tea == null) throw new NotFoundException();
+
             await _teaRepository.Delete(tea);
         }
